+++ bot.py
@@ -1,4 +1,4 @@
-__version__ = "3.7.14-dev5"
+__version__ = "3.7.14-dev6"
 
 
 import asyncio