+++ cogs/modmail.py
@@ -1050,8 +1050,30 @@ class Modmail(commands.Cog):
 
         roles = []
         users = []
+        now = ctx.message.created_at
 
-        for id_, reason in self.bot.blocked_users.items():
+        blocked_users = list(self.bot.blocked_users.items())
+        for id_, reason in blocked_users:
+            # parse "reason" and check if block is expired
+            # etc "blah blah blah... until 2019-10-14T21:12:45.559948."
+            end_time = re.search(r"until ([^`]+?)\.$", reason)
+            if end_time is None:
+                # backwards compat
+                end_time = re.search(r"%([^%]+?)%", reason)
+                if end_time is not None:
+                    logger.warning(
+                        r"Deprecated time message for user %s, block and unblock again to update.",
+                        id_,
+                    )
+
+            if end_time is not None:
+                after = (datetime.fromisoformat(end_time.group(1)) - now).total_seconds()
+                if after <= 0:
+                    # No longer blocked
+                    self.bot.blocked_users.pop(str(id_))
+                    logger.debug("No longer blocked, user %s.", id_)
+                    continue
+
             user = self.bot.get_user(int(id_))
             if user:
                 users.append((user.mention, reason))
@@ -1062,7 +1084,28 @@ class Modmail(commands.Cog):
                 except discord.NotFound:
                     users.append((id_, reason))
 
-        for id_, reason in self.bot.blocked_roles.items():
+        blocked_roles = list(self.bot.blocked_roles.items())
+        for id_, reason in blocked_roles:
+            # parse "reason" and check if block is expired
+            # etc "blah blah blah... until 2019-10-14T21:12:45.559948."
+            end_time = re.search(r"until ([^`]+?)\.$", reason)
+            if end_time is None:
+                # backwards compat
+                end_time = re.search(r"%([^%]+?)%", reason)
+                if end_time is not None:
+                    logger.warning(
+                        r"Deprecated time message for role %s, block and unblock again to update.",
+                        id_,
+                    )
+
+            if end_time is not None:
+                after = (datetime.fromisoformat(end_time.group(1)) - now).total_seconds()
+                if after <= 0:
+                    # No longer blocked
+                    self.bot.blocked_roles.pop(str(id_))
+                    logger.debug("No longer blocked, role %s.", id_)
+                    continue
+
             role = self.bot.guild.get_role(int(id_))
             if role:
                 roles.append((role.mention, reason))
@@ -1175,7 +1218,7 @@ class Modmail(commands.Cog):
         after: UserFriendlyTime = None,
     ):
         """
-        Block a user from using Modmail.
+        Block a user or role from using Modmail.
 
         You may choose to set a time as to when the user will automatically be unblocked.
 
@@ -1190,9 +1233,9 @@ class Modmail(commands.Cog):
             if thread:
                 user_or_role = thread.recipient
             elif after is None:
-                raise commands.MissingRequiredArgument(SimpleNamespace(name="user"))
+                raise commands.MissingRequiredArgument(SimpleNamespace(name="user or role"))
             else:
-                raise commands.BadArgument(f'User "{after.arg}" not found.')
+                raise commands.BadArgument(f'User or role "{after.arg}" not found.')
 
         mention = getattr(user_or_role, "mention", f"`{user_or_role.id}`")